refactor: split DetectionMutationTests into one file per nested suite
